chore: drop v2 naming on client session
