+++ tests/client/session/test_elicitation_handling.py
@@ -1,7 +1,7 @@
 from unittest.mock import AsyncMock
 
 from conduit.client.protocol.elicitation import ElicitationNotConfiguredError
-from conduit.client.session_v2 import ClientConfig, ClientSession
+from conduit.client.session import ClientConfig, ClientSession
 from conduit.protocol.base import INTERNAL_ERROR, METHOD_NOT_FOUND, Error
 from conduit.protocol.elicitation import ElicitRequest, ElicitResult
 from conduit.protocol.initialization import ClientCapabilities, Implementation
